% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bipw_priors}
\alias{bipw_priors}
\title{Collect priors for both outcome and propensity models (deprecated)}
\usage{
bipw_priors(outcome = smoothbp_priors(), propensity = prior_normal(0, 2.5))
}
\description{
Deprecated: use \code{\link{bjlm_priors}} instead.
}
