% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{bjlm_model}
\alias{bjlm_model}
\title{Initialise a Bayesian Joint Longitudinal Model (BJLM)}
\usage{
bjlm_model()
}
\value{
A \code{bjlm_model} object.
}
\description{
Initialise a Bayesian Joint Longitudinal Model (BJLM)
}
