% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bjlm_outcome_priors}
\alias{bjlm_outcome_priors}
\title{Alias for smoothbp_priors}
\usage{
bjlm_outcome_priors(
  b0 = prior_normal(0, 10),
  b1 = prior_normal(0, 2),
  deltas = prior_normal(0, 2),
  omega = prior_normal(3, 2, lb = 0),
  rho = prior_normal(3, 2, lb = 0),
  sigma = prior_invgamma(1, 1),
  sigma_u = prior_invgamma(1, 1),
  sigma_re_om = prior_invgamma(1, 1)
)
}
\description{
Alias for smoothbp_priors
}
