% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bjlm_priors}
\alias{bjlm_priors}
\title{Collect priors for both outcome and propensity models in a joint BJLM model}
\usage{
bjlm_priors(outcome = smoothbp_priors(), propensity = prior_normal(0, 2.5))
}
\arguments{
\item{outcome}{An object of class \code{smoothbp_priors} specifying priors
for the outcome model. See \code{\link{smoothbp_priors}} for details.}

\item{propensity}{An object of class \code{smoothbp_prior} (of normal family)
specifying the prior for the propensity coefficients. Default is \code{prior_normal(0, 2.5)}.}
}
\value{
A \code{bjlm_priors} list.
}
\description{
Collect priors for both outcome and propensity models in a joint BJLM model
}
