% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{compile}
\alias{compile}
\title{Compile a BJLM model}
\usage{
compile(model)
}
\arguments{
\item{model}{A \code{bjlm_model} object.}
}
\value{
A \code{bjlm_compiled_model} object ready to be fitted.
}
\description{
Performs key alignment, demographic matching, strict Bayesian Cut validation,
and validates all GP specifications.
}
