% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{.compute_oos_log_lik}
\alias{.compute_oos_log_lik}
\title{Helper to compute out-of-sample pointwise log-likelihood matrix}
\usage{
.compute_oos_log_lik(fit_active, validation_data, y_name)
}
\arguments{
\item{fit_active}{A fitted \code{bjlm_fit} model.}

\item{validation_data}{A data frame containing held-out observations.}

\item{y_name}{Character. Name of response variable.}
}
\value{
An S x N matrix of log-likelihoods.
}
\description{
Helper to compute out-of-sample pointwise log-likelihood matrix
}
\keyword{internal}
