% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{.reconstruct_model}
\alias{.reconstruct_model}
\title{Reconstruct a bjlm_model specification from a bjlm_fit object}
\usage{
.reconstruct_model(object, data)
}
\arguments{
\item{object}{A \code{bjlm_fit} object.}

\item{data}{A data frame containing the subset of data for refitting.}
}
\value{
A \code{bjlm_model} specification.
}
\description{
Helper function for LFO-CV refitting.
}
\keyword{internal}
