% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{fit.bjlm_compiled_model}
\alias{fit.bjlm_compiled_model}
\title{Fit a compiled BJLM model}
\usage{
\method{fit}{bjlm_compiled_model}(object, priors = NULL, ...)
}
\arguments{
\item{object}{A \code{bjlm_compiled_model} object.}

\item{priors}{A \code{bjlm_priors} object specifying priors.}

\item{...}{Additional arguments passed to the fitting engine \code{bjlm()}.}
}
\value{
A \code{bjlm_fit} object.
}
\description{
Fit a compiled BJLM model
}
