% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{fit}
\alias{fit}
\title{Fit a compiled model}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A compiled model object.}

\item{...}{Additional arguments.}
}
\value{
A fitted model object.
}
\description{
Fit a compiled model
}
