% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{fixed}
\alias{fixed}
\alias{prior_fixed}
\title{Fix a parameter at a specific value}
\usage{
fixed(value)

prior_fixed(value)
}
\arguments{
\item{value}{The fixed value(s) (numeric scalar or vector).}
}
\value{
A \code{smoothbp_fixed} object.
}
\description{
Used within \code{omega} or \code{rho} lists in \code{\link{bjlm}} to specify that a
parameter is fixed and should not be estimated.
}
