% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{latent_gp}
\alias{latent_gp}
\title{Add a latent Gaussian Process (GP) confounder to a BJLM model}
\usage{
latent_gp(model, name, time, subject, kernel = "se")
}
\arguments{
\item{model}{A \code{bjlm_model} object.}

\item{name}{Character name of the GP confounder.}

\item{time}{Name of the time variable in the outcome dataset.}

\item{subject}{Name of the subject grouping variable.}

\item{kernel}{Kernel function to use. Currently supports \code{"se"} (Squared Exponential).}
}
\value{
The modified \code{bjlm_model} object.
}
\description{
Add a latent Gaussian Process (GP) confounder to a BJLM model
}
