% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{log_lik.bjlm_fit}
\alias{log_lik.bjlm_fit}
\title{Extract pointwise log-likelihood from a bjlm_fit object}
\usage{
\method{log_lik}{bjlm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{bjlm_fit} object.}

\item{...}{Additional arguments passed to \code{fitted}.}
}
\value{
A matrix of size S x N containing pointwise log-likelihoods.
}
\description{
Extract pointwise log-likelihood from a bjlm_fit object
}
