% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{log_lik}
\alias{log_lik}
\title{Pointwise log-likelihood matrix}
\usage{
log_lik(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Pointwise log-likelihood matrix
}
