% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.bjlm_fit}
\alias{loo.bjlm_fit}
\title{Compute LOO-IC for bjlm_fit objects}
\usage{
\method{loo}{bjlm_fit}(x, ...)
}
\arguments{
\item{x}{A \code{bjlm_fit} object.}

\item{...}{Additional arguments passed to \code{loo::loo}.}
}
\description{
Compute LOO-IC for bjlm_fit objects
}
