% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.smoothbp_fit}
\alias{loo.smoothbp_fit}
\title{Compute LOO-IC for smoothbp_fit objects}
\usage{
\method{loo}{smoothbp_fit}(x, ...)
}
\arguments{
\item{x}{A \code{smoothbp_fit} object.}

\item{...}{Additional arguments passed to \code{loo::loo}.}
}
\description{
Compute LOO-IC for smoothbp_fit objects
}
