% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{plot.bjlm_lfo}
\alias{plot.bjlm_lfo}
\title{Plot LFO-CV Diagnostics}
\usage{
\method{plot}{bjlm_lfo}(x, ...)
}
\arguments{
\item{x}{A \code{bjlm_lfo} object.}

\item{...}{Unused.}
}
\description{
Premium visualization of the Pareto k diagnostics and refit events over the time grid.
}
