% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_bipw.R
\name{print.bipw_fit}
\alias{print.bipw_fit}
\title{Print method for bipw_fit}
\usage{
\method{print}{bipw_fit}(x, ...)
}
\arguments{
\item{x}{A \code{bipw_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bipw_fit
}
