% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_bjlm.R
\name{print.bjlm_fit}
\alias{print.bjlm_fit}
\title{Print method for bjlm_fit}
\usage{
\method{print}{bjlm_fit}(x, ...)
}
\arguments{
\item{x}{A \code{bjlm_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bjlm_fit
}
