% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_halfnormal}
\alias{prior_halfnormal}
\title{Specify a half-normal prior for a scale (standard-deviation) parameter}
\usage{
prior_halfnormal(sd = 1)
}
\arguments{
\item{sd}{Scale of the half-normal (> 0); larger is more diffuse.}
}
\value{
A \code{smoothbp_prior} object.
}
\description{
Specify a half-normal prior for a scale (standard-deviation) parameter
}
