% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{propensity}
\alias{propensity}
\title{Add a propensity score block to a BJLM model}
\usage{
propensity(model, formula, data = NULL, family = binomial("logit"))
}
\arguments{
\item{model}{A \code{bjlm_model} object.}

\item{formula}{A two-sided formula of the form \code{treatment ~ covariates},
specifying the propensity model. The LHS must be a binary (0/1) variable.}

\item{data}{A data frame containing propensity score covariates.}

\item{family}{A description of the error distribution and link function.
Currently supports \code{binomial("logit")} (default).}
}
\value{
The modified \code{bjlm_model} object.
}
\description{
Add a propensity score block to a BJLM model
}
