% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loo}
\alias{waic}
\alias{pp_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayesplot}{\code{\link[bayesplot]{pp_check}}}

  \item{loo}{\code{\link[loo]{loo}}, \code{\link[loo]{waic}}}
}}

