% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_bjlm.R
\name{summary.bjlm_fit}
\alias{summary.bjlm_fit}
\title{Summary for bjlm_fit objects}
\usage{
\method{summary}{bjlm_fit}(object, model = c("both", "outcome", "propensity"), prob = 0.95, ...)
}
\arguments{
\item{object}{A \code{bjlm_fit} object.}

\item{model}{Which model to summarise: \code{"outcome"}, \code{"propensity"},
or \code{"both"} (default).}

\item{prob}{Probability mass for credible interval (default: 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary for bjlm_fit objects
}
