% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{waic.bjlm_fit}
\alias{waic.bjlm_fit}
\title{Compute WAIC for bjlm_fit objects}
\usage{
\method{waic}{bjlm_fit}(x, ...)
}
\arguments{
\item{x}{A \code{bjlm_fit} object.}

\item{...}{Additional arguments passed to \code{loo::waic}.}
}
\description{
Compute WAIC for bjlm_fit objects
}
