% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{waic.smoothbp_fit}
\alias{waic.smoothbp_fit}
\title{Compute WAIC for smoothbp_fit objects}
\usage{
\method{waic}{smoothbp_fit}(x, ...)
}
\arguments{
\item{x}{A \code{smoothbp_fit} object.}

\item{...}{Additional arguments passed to \code{loo::waic}.}
}
\description{
Compute WAIC for smoothbp_fit objects
}
