% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bjlm_outcome_priors}
\alias{bjlm_outcome_priors}
\title{Alias for smoothbp_priors}
\usage{
bjlm_outcome_priors(
  b0 = prior_normal(0, 10),
  b1 = prior_normal(0, 2),
  deltas = prior_normal(0, 2),
  omega = prior_normal(3, 2, lb = 0),
  rho = prior_normal(3, 2, lb = 0),
  sigma = prior_invgamma(1, 1),
  sigma_u = prior_invgamma(1, 1),
  sigma_re_om = prior_invgamma(1, 1),
  r = prior_gamma(1, 1)
)
}
\arguments{
\item{b0}{Prior(s) for \code{b0} regression coefficients.}

\item{b1}{Prior(s) for \code{b1} regression coefficients.}

\item{deltas}{Prior(s) for slope change coefficients (one list per segment).}

\item{omega}{Prior(s) for \code{omega} coefficients (one list per segment).}

\item{rho}{Prior(s) for \code{rho} coefficients (one list per segment).}

\item{sigma}{\code{prior_invgamma()} for residual SD.}

\item{sigma_u}{\code{prior_invgamma()} for random-effect SD.}

\item{sigma_re_om}{\code{prior_invgamma()} for random-effect SD on omega.}

\item{r}{\code{prior_gamma()} for Negative Binomial overdispersion parameter.}
}
\description{
Alias for smoothbp_priors
}
