% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_api.R
\name{latent_gp}
\alias{latent_gp}
\title{Add a latent Gaussian Process (GP) confounder to a BJLM model}
\usage{
latent_gp(
  model,
  name,
  data,
  obs_var,
  time_var,
  time_trt_var,
  time_out_var,
  subject,
  kernel = "se",
  priors = gp_priors()
)
}
\arguments{
\item{model}{A \code{bjlm_model} object.}

\item{name}{Character name of the GP confounder.}

\item{data}{A data frame containing the observed covariate data.}

\item{obs_var}{The column name of the noisy continuous observations in \code{data}.}

\item{time_var}{The column name of the time variable in \code{data}.}

\item{time_trt_var}{The column name of the time variable in the propensity data.}

\item{time_out_var}{The column name of the time variable in the outcome data.}

\item{subject}{Name of the subject grouping variable.}

\item{kernel}{Kernel function to use. Currently supports \code{"se"} (Squared Exponential).}

\item{priors}{A \code{\link[=gp_priors]{gp_priors()}} bundle for the GP hyperparameters (marginal SD
\code{alpha}, lengthscale \code{rho}, observation-noise SD \code{sigma_x}). Defaults to
\code{gp_priors()}, i.e. lognormal(0,1) on \code{alpha}, the resolution-aware lengthscale
prior on \code{rho}, and lognormal(-1,1) on \code{sigma_x}.}
}
\value{
The modified \code{bjlm_model} object.
}
\description{
Add a latent Gaussian Process (GP) confounder to a BJLM model
}
