% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot time-varying latent Gaussian Process (GP) deviations}
\usage{
plot_gp(fit, subjects = NULL, n_subjects = 5, ...)
}
\arguments{
\item{fit}{A \code{bjlm_fit} object.}

\item{subjects}{Optional vector of subject IDs to plot.}

\item{n_subjects}{Number of random subjects to select if \code{subjects} is NULL.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot time-varying latent Gaussian Process (GP) deviations
}
